+++ src/confd/src/interfaces.h
@@ -130,6 +130,7 @@ int bridge_port_gen(struct lyd_node *dif, struct lyd_node *cif, FILE *ip);
 typedef enum wifi_mode_t {
 	wifi_station,
 	wifi_ap,
+	wifi_mesh,
 	wifi_unknown
 } wifi_mode_t;
 
@@ -138,6 +139,7 @@ int wifi_add_iface(struct lyd_node *cif, struct dagger *net);
 int wifi_del_iface(struct lyd_node *dif, struct dagger *net);
 int wifi_mode_changed(struct lyd_node *wifi);
 int wifi_gen_station(struct lyd_node *cif);
+int wifi_gen_mesh(struct lyd_node *cif);
 wifi_mode_t wifi_get_mode(struct lyd_node *wifi);
 
 /* if-gre.c */